requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
bcrypt==4.2.1
python-jose[cryptography]==3.3.0